MANY UPDATE MUCH WOW!!!
I have added a new lectures and assignment page to allow access to all
of the lecture and assignment pages that will be linked through the
calendar as time moves on. I have also added all of the appropriate
files for these pages, and most of the assignment pages are completed.